+++ R/get_stats.R
@@ -1,7 +1,7 @@
 #' A pipe friendly way to get summary stats for exploratory data analysis
 #' @param x The data for which stats are required
 #' @param func The nature of function to apply
-#' @param exclude What kind of data should be excluded? Defaults to c("character","factor") to drop character and factor columns
+#' @param exclude What kind of data should be excluded? Use for example c("character","factor") to drop character and factor columns
 #' @param na.rm Logical. Should NAs be removed. Defaults to FALSE.
 #' @param na_action If na.rm is set to TRUE, this uses na_replace to replace missing values.
 #' @param ... Other arguments to na_replace
@@ -11,13 +11,13 @@
 #' @examples
 #' head(get_data_Stats(airquality,mean,na.rm = TRUE,na_action = "get_mode"))
 #' @export 
-get_data_Stats<-function (x=NULL, func=NULL,exclude=c("factor","character"),na.rm=FALSE,na_action=NULL,...) {
+get_data_Stats<-function (x=NULL, func=NULL,exclude=NULL,na.rm=FALSE,na_action=NULL,...) {
   UseMethod("get_data_Stats")
 }
 
 #' @export
 
-get_data_Stats.data.frame<-function (x=NULL, func=NULL,exclude=c("factor","character"),na.rm=FALSE,na_action=NULL,...) {
+get_data_Stats.data.frame<-function (x=NULL, func=NULL,exclude=NULL,na.rm=FALSE,na_action=NULL,...) {
 
 if(any(is.null(x), is.null(func))) stop("Both df and func must be supplied")